+++ base/p_mobj.c
@@ -523,7 +523,7 @@ static void P_XYMovement(mobj_t *mo)
 					{
 					case MT_CENTAUR:
 					case MT_CENTAURLEADER:
-						if (abs(angle - BlockingMobj->angle)>>24 > 45)
+						if (abs((int)angle - (int)BlockingMobj->angle)>>24 > 45)
 							goto explode;
 						if (mo->type == MT_HOLY_FX)
 							goto explode;
